feat(architecture): separate route logic from controller
