+++ api/controllers/auth/index.ts
@@ -1,0 +1,4 @@
+export { login } from './login';
+export { logout } from './logout';
+export { refreshToken } from './refreshToken';
+export { register } from './register';
