+++ api/controllers/auth/logout.test.ts
