+++ api/controllers/auth/register.test.ts
