+++ api/controllers/events/index.ts
@@ -1,0 +1,9 @@
+export {
+  getAllEvents,
+  getEventById,
+  createEvent,
+  modifyEventById,
+  deleteEventById,
+} from './events';
+
+export { addAdminToEvent } from './admin';
