+++ api/controllers/groups/index.ts
@@ -1,0 +1,7 @@
+export {
+  getGroupById,
+  getAllGroup,
+  createGroup,
+  modifyGroupById,
+  deleteGroupById,
+} from './groups';
