+++ api/controllers/info/index.ts
@@ -1,0 +1,1 @@
+export { getInfo } from './info';
