+++ api/controllers/instruments/index.ts
@@ -1,0 +1,1 @@
+export { getInstruments } from './instruments';
