+++ api/controllers/musicians/index.ts
@@ -1,0 +1,1 @@
+export { getAllMusicians, getMusicianById } from './musicians';
