+++ api/controllers/profil/index.ts
@@ -1,0 +1,1 @@
+export { getUserProfil, modifyUserProfil, deleteUserProfil } from './profil';
