+++ api/routes/info/index.ts
@@ -1,0 +1,8 @@
+import express from 'express';
+import * as infoController from '../../controllers/info';
+
+const router = express.Router();
+
+router.get('/', infoController.getInfo);
+
+export default router;
