+++ api/routes/instruments/index.ts
@@ -1,0 +1,8 @@
+import express from 'express';
+import { getInstruments } from '../../controllers/instruments/instruments';
+
+const router = express.Router();
+
+router.get('/', getInstruments);
+
+export default router;
