+++ api/routes/profil/index.ts
@@ -1,0 +1,12 @@
+import express from 'express';
+import * as profilController from '../../controllers/profil';
+
+const router = express.Router();
+
+router.get('/', profilController.getUserProfil);
+
+router.patch('/', profilController.modifyUserProfil);
+
+router.delete('/', profilController.deleteUserProfil);
+
+export default router;
